+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 - All assertion methods now return the checked value.
 - Added `notInArray` and `notOneOf`.
 - Added `isInitialized`, to check if a class property is initialized.
+- Added `notNegativeInteger` and `negativeInteger`
 
 ### Fixed